Node representation changed
instead of having an unit in charge of when to add new representation of the list/queue/stack, now the nodes themselves tells when to create one.
works as before (BFS/DFS OK), but Dijkstra still have some issues updating the representation according to the values of the nodes (when updating values of list elements occure).